Resue EventName in internal listeners instead of EventName type alias, move to its own file
