+++ crates/tauri-cli/src/helpers/flock.rs
@@ -327,15 +327,11 @@ mod sys {
   }
 
   pub(super) fn error_contended(err: &Error) -> bool {
-    err
-      .raw_os_error()
-      .map_or(false, |x| x == ERROR_LOCK_VIOLATION as i32)
+    err.raw_os_error() == Some(ERROR_LOCK_VIOLATION as i32)
   }
 
   pub(super) fn error_unsupported(err: &Error) -> bool {
-    err
-      .raw_os_error()
-      .map_or(false, |x| x == ERROR_INVALID_FUNCTION as i32)
+    err.raw_os_error() == Some(ERROR_INVALID_FUNCTION as i32)
   }
 
   pub(super) fn unlock(file: &File) -> Result<()> {
