+++ crates/tauri/src/manager/mod.rs
@@ -509,7 +509,7 @@ impl<R: Runtime> AppManager<R> {
 
   /// # Panics
   /// Will panic if `event` contains characters other than alphanumeric, `-`, `/`, `:` and `_`
-  pub fn listen<F: Fn(Event) + Send + 'static>(
+  pub(crate) fn listen<F: Fn(Event) + Send + 'static>(
     &self,
     event: EventName,
     target: EventTarget,
@@ -520,7 +520,7 @@ impl<R: Runtime> AppManager<R> {
 
   /// # Panics
   /// Will panic if `event` contains characters other than alphanumeric, `-`, `/`, `:` and `_`
-  pub fn once<F: FnOnce(Event) + Send + 'static>(
+  pub(crate) fn once<F: FnOnce(Event) + Send + 'static>(
     &self,
     event: EventName,
     target: EventTarget,
@@ -537,7 +537,11 @@ impl<R: Runtime> AppManager<R> {
     feature = "tracing",
     tracing::instrument("app::emit", skip(self, payload))
   )]
-  pub fn emit<S: Serialize>(&self, event: EventName<&str>, payload: &S) -> crate::Result<()> {
+  pub(crate) fn emit<S: Serialize>(
+    &self,
+    event: EventName<&str>,
+    payload: &S,
+  ) -> crate::Result<()> {
     #[cfg(feature = "tracing")]
     let _span = tracing::debug_span!("emit::run").entered();
     let emit_args = EmitArgs::new(event, &payload)?;
@@ -560,7 +564,7 @@ impl<R: Runtime> AppManager<R> {
     feature = "tracing",
     tracing::instrument("app::emit::filter", skip(self, payload, filter))
   )]
-  pub fn emit_filter<S, F>(
+  pub(crate) fn emit_filter<S, F>(
     &self,
     event: EventName<&str>,
     payload: S,
@@ -591,7 +595,12 @@ impl<R: Runtime> AppManager<R> {
     feature = "tracing",
     tracing::instrument("app::emit::to", skip(self, target, payload), fields(target))
   )]
-  pub fn emit_to<I, S>(&self, target: I, event: EventName<&str>, payload: &S) -> crate::Result<()>
+  pub(crate) fn emit_to<I, S>(
+    &self,
+    target: I,
+    event: EventName<&str>,
+    payload: &S,
+  ) -> crate::Result<()>
   where
     I: Into<EventTarget>,
     S: Serialize,
